moved DocumentTools into gui project
