+++ WolvenKit.Common/Services/HashService.cs
@@ -205,10 +205,6 @@ namespace WolvenKit.Common.Services
             }
             TweakDBIDPool.SetNative(dict);
         }
-
-        public virtual bool AddResourcePath(string resourcePath) => false;
-
-        public virtual bool AddTweakName(string tweakName) => false;
         
         #endregion Methods
     }
