+++ WolvenKit/Views/Documents/RedDocumentViewMenuBar.xaml.cs
@@ -26,7 +26,6 @@ using WolvenKit.Core.Exceptions;
 using WolvenKit.Core.Interfaces;
 using WolvenKit.Core.Services;
 using WolvenKit.Helpers;
-using WolvenKit.Modkit.RED4;
 using WolvenKit.Modkit.RED4.Tools;
 using WolvenKit.RED4.Types;
 using WolvenKit.Views.Dialogs.Windows;
